Rename ai types folder to lighthouse
