+++ ui/app/(ai)/analyst/(tools)/resources.ts
@@ -1,7 +1,7 @@
 import { tool } from "@langchain/core/tools";
 
 import { aiGetResource, aiGetResources } from "@/actions/lighthouse/resources";
-import { getResourceSchema, getResourcesSchema } from "@/types/ai";
+import { getResourceSchema, getResourcesSchema } from "@/types/lighthouse";
 
 export const getResourcesTool = tool(
   async ({ page, query, sort, filters, fields }) => {
