+++ ui/types/lighthouse/index.ts
