+++ ui/types/lighthouse/resources.ts
