+++ ui/types/lighthouse/scans.ts
